define new structure project
