+++ api/routes.py
@@ -1,0 +1,77 @@
+from datetime import timedelta
+from fastapi import APIRouter, Depends, HTTPException, status
+from sqlalchemy.orm import Session
+from schemas.user import UserCreate, User
+from schemas.auth import Token
+from services.auth import (
+    create_user,
+    authenticate_user,
+    create_access_token,
+    ACCESS_TOKEN_EXPIRE_MINUTES,
+    get_user,
+    get_current_user,
+    get_current_active_user
+)
+from database import get_db
+from fastapi.security import OAuth2PasswordRequestForm
+
+
+router = APIRouter()
+
+@router.post("/token", response_model=Token)
+async def login_for_access_token(
+    form_data: OAuth2PasswordRequestForm = Depends(),
+    db: Session = Depends(get_db)
+):
+    user = authenticate_user(db, form_data.username, form_data.password)
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Incorrect username or password",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+    access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
+    access_token = create_access_token(
+        data={"sub": user.username},
+        expires_delta=access_token_expires
+    )
+    return {"access_token": access_token, "token_type": "bearer"}
+
+@router.post("/users/", response_model=User)
+async def create_new_user(
+    user: UserCreate,
+    db: Session = Depends(get_db)
+):
+    db_user = get_user(db, username=user.username)
+    if db_user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Username already registered"
+        )
+    return create_user(db=db, user=user)
+
+@router.get("/users/me", response_model=User)
+async def read_users_me(current_user: User = Depends(get_current_user)):
+
+    return current_user
+
+
+@router.get("/users/{username}", response_model=User)
+async def read_user(
+    username: str,
+    current_user: User = Depends(get_current_user),
+    db: Session = Depends(get_db)
+):
+    db_user = get_user(db, username=username)
+    if db_user is None:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="User not found"
+        )
+    return db_user
+
+@router.get("/users/me/validate-token")
+async def validate_token(
+    current_user: User = Depends(get_current_active_user)
+):
+    return {"valid": True, "username": current_user.username}
