+++ models/__init__.py
@@ -1,0 +1,2 @@
+from database import Base
+from .user import UserDB
