+++ models/user.py
@@ -1,0 +1,12 @@
+from sqlalchemy import Column, Integer, String, Boolean
+from database import Base
+
+class UserDB(Base):
+    __tablename__ = "users"
+
+    id = Column(Integer, primary_key=True, index=True)
+    username = Column(String, unique=True, index=True)
+    email = Column(String, unique=True, index=True, nullable=True)
+    full_name = Column(String, nullable=True)
+    hashed_password = Column(String)
+    disabled = Column(Boolean, default=False)
