+++ requirements.txt
@@ -1,8 +1,47 @@
-fastapi
-uvicorn[standard]
-python-jose[cryptography]
-passlib[bcrypt]
-psycopg2-binary
-PyJWT
-sqlalchemy
-python-multipart
+fastapi~=0.115.3
+uvicorn[standard]~=0.32.0
+python-jose[cryptography]~=3.3.0
+passlib[bcrypt]~=1.7.4
+psycopg2-binary~=2.9.10
+PyJWT~=2.9.0
+sqlalchemy~=2.0.36
+python-multipart~=0.0.12
+dnspython~=2.7.0
+idna~=3.10
+httpcore~=1.0.6
+httpx~=0.27.2
+cryptography~=43.0.3
+anyio~=4.6.2.post1
+h11~=0.14.0
+pip~=24.2
+rsa~=4.9
+pyasn1~=0.6.1
+cffi~=1.17.1
+setuptools~=59.6.0
+ecdsa~=0.19.0
+rich~=13.9.3
+Pygments~=2.18.0
+markdown-it-py~=3.0.0
+PyYAML~=6.0.2
+sniffio~=1.3.1
+typing_extensions~=4.12.2
+six~=1.16.0
+click~=8.1.7
+certifi~=2024.8.30
+mdurl~=0.1.2
+typer~=0.12.5
+shellingham~=1.5.4
+MarkupSafe~=3.0.2
+uvloop~=0.21.0
+starlette~=0.41.0
+email_validator~=2.2.0
+pydantic~=2.9.2
+pydantic_core~=2.23.4
+websockets~=13.1
+httptools~=0.6.4
+watchfiles~=0.24.0
+greenlet~=3.1.1
+annotated-types~=0.7.0
+pycparser~=2.22
+Jinja2~=3.1.4
+bcrypt~=3.1.0
