+++ schemas/auth.py
@@ -1,0 +1,6 @@
+# schemas.py
+from pydantic import BaseModel
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
